requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
email-validator==2.2.0
psycopg[binary]==3.2.5
gunicorn==23.0.0
python-dotenv==1.0.1
pytest==8.3.4
Werkzeug==3.1.3
tzdata==2025.2
